pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "murimsim"
version = "0.1.0"
description = "Multi-agent martial arts sect simulation"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pyyaml>=6.0",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]
rl = [
    "torch>=2.1",
    "gymnasium>=0.29",
    "stable-baselines3>=2.2",
    "pettingzoo>=1.24",
    "wandb>=0.16",
]
llm = [
    "litellm>=1.0",
]
viz = [
    "pygame>=2.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["murimsim*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["murimsim"]
